requirements.txt
certifi==2024.8.30
charset-normalizer==3.3.2
idna==3.8
lxml==5.3.0
PyAudio==0.2.14
pyspellchecker==0.8.1
python-docx==1.1.2
requests==2.32.3
setuptools==74.0.0
SpeechRecognition==3.10.4
typing_extensions==4.12.2
urllib3==2.2.2
